% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_location.R
\name{cache_location_ids}
\alias{cache_location_ids}
\title{Cache Location IDs to System Cache}
\usage{
cache_location_ids(api_key = get_api_key(), refresh_cache = FALSE)
}
\arguments{
\item{api_key}{A character string containing your API key.}

\item{refresh_cache}{Logical indicating whether to refresh the cache even
if an RDS file exists.}
}
\value{
Invisibly returns the vector of location IDs.
}
\description{
Retrieves all location data from the CQC API (using scrape_locations()),
extracts the location IDs, and saves them to the system cache directory.
}
