% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_provider.R
\name{cache_provider_ids}
\alias{cache_provider_ids}
\title{Cache Provider IDs to System Cache}
\usage{
cache_provider_ids(api_key = get_api_key(), refresh_cache = FALSE)
}
\arguments{
\item{api_key}{A character string containing your API key.}

\item{refresh_cache}{Logical indicating whether to refresh the cache even
if an RDS file exists.}
}
\value{
Invisibly returns the vector of provider IDs.
}
\description{
Retrieves all provider data from the CQC API (using scrape_providers()),
extracts the provider IDs, and saves them to the system cache directory.
}
