% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes_utils.R
\name{get_cqc_changes}
\alias{get_cqc_changes}
\title{Get CQC Changes and Cache to System Directory}
\usage{
get_cqc_changes(
  startDate,
  endDate,
  api_key = get_api_key(),
  refresh_cache = FALSE
)
}
\arguments{
\item{startDate}{Start date in DD.MM.YYYY format}

\item{endDate}{End date in DD.MM.YYYY format}

\item{api_key}{API key (defaults to environment variable)}

\item{refresh_cache}{Logical indicating whether to refresh cache even if file exists}
}
\value{
Data frame with location_changes and provider_changes columns
}
\description{
Get CQC Changes and Cache to System Directory
}
