% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes_utils.R
\name{get_incremental_changes}
\alias{get_incremental_changes}
\title{Get Incremental Changes and Cache to System Directory}
\usage{
get_incremental_changes(
  dataset_file = "location_df.rda",
  data_folder = "data",
  date_column = "folder_date",
  api_key = get_api_key()
)
}
\arguments{
\item{dataset_file}{Dataset file name (default: "location_df.rda")}

\item{data_folder}{Path to data folder (default: "data")}

\item{date_column}{Column containing date information (default: "folder_date")}

\item{api_key}{API key (defaults to get_api_key())}
}
\value{
Data frame with location_changes and provider_changes columns
}
\description{
Get Incremental Changes and Cache to System Directory
}
