% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{get_paginated_data}
\alias{get_paginated_data}
\title{Get paginated data from the CQC API}
\usage{
get_paginated_data(base_url, headers, query_params = list(), per_page = 100)
}
\arguments{
\item{base_url}{The base URL of the API endpoint. Uses 'Get Location By ID'
endpoint}

\item{headers}{A list of HTTP headers for the request.}

\item{query_params}{A named list of additional query parameters.}

\item{per_page}{Number of records per page (default is 100).}
}
\value{
A data frame containing combined data from all pages.
}
\description{
Retrieves data from a paginated endpoint of the CQC API.
}
