% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_merger.R
\name{merge_provider_location}
\alias{merge_provider_location}
\title{Merge Provider and Location Data and Save to Package}
\usage{
merge_provider_location(join_fn = dplyr::inner_join)
}
\arguments{
\item{join_fn}{Join function to use (default: dplyr::inner_join)}
}
\value{
Invisibly returns the merged data frame
}
\description{
Merge Provider and Location Data and Save to Package
}
