% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_location.R
\name{scrape_locations}
\alias{scrape_locations}
\title{Scrape Locations from the CQC API}
\usage{
scrape_locations(api_key, per_page = 100, verbose = FALSE)
}
\arguments{
\item{api_key}{A character string containing your API key.}

\item{per_page}{Number of records per page.}

\item{verbose}{Logical; if TRUE, prints number of location IDs retrieved.}
}
\value{
A data frame of location data.
}
\description{
Retrieves all location data from the CQC API
using the paginated locations endpoint.
}
