% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_provider.R
\name{scrape_providers}
\alias{scrape_providers}
\title{Scrape Providers from the CQC API}
\usage{
scrape_providers(api_key, per_page = 100, verbose = FALSE)
}
\arguments{
\item{api_key}{A character string containing your API key.}

\item{per_page}{Number of records per page.}

\item{verbose}{Logical; if TRUE, prints number of provider IDs retrieved.}
}
\value{
A data frame of providers data.
}
\description{
Retrieves all provider data from the CQC API
using the paginated provider endpoint.
}
