% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{set_headers}
\alias{set_headers}
\title{Set HTTP Headers for CQC API}
\usage{
set_headers(api_key = NULL)
}
\arguments{
\item{api_key}{Optional character string containing the CQC API subscription
key. If not provided, it will be retrieved using `get_api_key()`.}
}
\value{
httr headers object that can be used in HTTP requests
}
\description{
Creates the appropriate HTTP headers for authenticating with the CQC API.
}
