% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changes_provider.R
\name{update_provider_dataset}
\alias{update_provider_dataset}
\title{Update Provider Dataset with Incremental Changes}
\usage{
update_provider_dataset(
  data_folder = "data",
  dataset_file = "provider_df.rda",
  date_column = "folder_date"
)
}
\arguments{
\item{data_folder}{Path to data folder (default: "data")}

\item{dataset_file}{Dataset file name (default: "provider_df.rda")}

\item{date_column}{Column containing date information (default: "folder_date")}
}
\value{
Invisibly returns summary information
}
\description{
Update Provider Dataset with Incremental Changes
}
